backend/requirements.txt
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
coreapi==2.3.3